Refactor transformers and generalise size transformer
