+++ modules/AppearancePluginUI/src/main/java/org/gephi/ui/appearance/plugin/color/PartitionColorTransformerPanel.form
