+++ modules/AppearancePluginUI/src/main/java/org/gephi/ui/appearance/plugin/size/RankingSizeTransformerPanel.form
