+++ modules/AppearancePluginUI/src/main/java/org/gephi/ui/appearance/plugin/unique/UniqueColorTransformerPanel.form
