+++ modules/AppearancePluginUI/src/main/java/org/gephi/ui/appearance/plugin/unique/UniqueElementColorTransformerUI.java
@@ -40,15 +40,15 @@
  Portions Copyrighted 2013 Gephi Consortium.
  */
 
-package org.gephi.ui.appearance.plugin;
+package org.gephi.ui.appearance.plugin.unique;
 
 import javax.swing.AbstractButton;
 import javax.swing.Icon;
 import javax.swing.JPanel;
 import org.gephi.appearance.api.Function;
-import org.gephi.appearance.api.PartitionFunction;
-import org.gephi.appearance.plugin.PartitionLabelColorTransformer;
-import org.gephi.appearance.spi.PartitionTransformer;
+import org.gephi.appearance.api.SimpleFunction;
+import org.gephi.appearance.plugin.UniqueElementColorTransformer;
+import org.gephi.appearance.spi.SimpleTransformer;
 import org.gephi.appearance.spi.TransformerCategory;
 import org.gephi.appearance.spi.TransformerUI;
 import org.gephi.ui.appearance.plugin.category.DefaultCategory;
@@ -58,24 +58,19 @@ import org.openide.util.lookup.ServiceProvider;
 /**
  * @author mbastian
  */
-@ServiceProvider(service = TransformerUI.class, position = 600)
-public class PartitionLabelColorTransformerUI implements TransformerUI {
+@ServiceProvider(service = TransformerUI.class, position = 100)
+public class UniqueElementColorTransformerUI implements TransformerUI {
 
-    private PartitionColorTransformerPanel panel;
-
-    @Override
-    public TransformerCategory getCategory() {
-        return DefaultCategory.LABEL_COLOR;
-    }
+    private UniqueColorTransformerPanel panel;
 
     @Override
     public String getDisplayName() {
-        return NbBundle.getMessage(PartitionLabelColorTransformerUI.class, "Attribute.partition.name");
+        return NbBundle.getMessage(UniqueElementColorTransformerUI.class, "Unique.name");
     }
 
     @Override
-    public Icon getIcon() {
-        return null;
+    public TransformerCategory getCategory() {
+        return DefaultCategory.COLOR;
     }
 
     @Override
@@ -84,24 +79,26 @@ public class PartitionLabelColorTransformerUI implements TransformerUI {
     }
 
     @Override
+    public Icon getIcon() {
+        return null;
+    }
+
+    @Override
     public synchronized JPanel getPanel(Function function) {
         if (panel == null) {
-            panel = new PartitionColorTransformerPanel();
+            panel = new UniqueColorTransformerPanel();
         }
-        panel.setup((PartitionFunction) function);
+        panel.setup((SimpleFunction) function);
         return panel;
     }
 
     @Override
     public synchronized AbstractButton[] getControlButton() {
-        if (panel == null) {
-            panel = new PartitionColorTransformerPanel();
-        }
-        return new AbstractButton[] {panel.getPaletteButton()};
+        return null;
     }
 
     @Override
-    public Class<? extends PartitionTransformer> getTransformerClass() {
-        return PartitionLabelColorTransformer.class;
+    public Class<? extends SimpleTransformer> getTransformerClass() {
+        return UniqueElementColorTransformer.class;
     }
 }
